extract storeMissingOptions into options module and get tests to pass
