--- add-on/src/lib/option-defaults.js
@@ -1,16 +1,0 @@
-'use strict'
-/* eslint-env browser, webextensions */
-
-module.exports = Object.freeze({
-  publicGatewayUrl: 'https://ipfs.io',
-  useCustomGateway: true,
-  automaticMode: true,
-  linkify: false,
-  dnslink: false,
-  preloadAtPublicGateway: true,
-  catchUnhandledProtocols: true,
-  displayNotifications: true,
-  customGatewayUrl: 'http://127.0.0.1:8080',
-  ipfsApiUrl: 'http://127.0.0.1:5001',
-  ipfsApiPollMs: 3000
-})
