+++ package.json
@@ -19,7 +19,7 @@
     "build:bundle-extension": "web-ext build -s add-on/ -i src/ -a build/",
     "watch": "run-p watch:*",
     "watch:js": "watchify add-on/src/background/background.js add-on/src/options/options.js add-on/src/popup/browser-action.js add-on/src/popup/quick-upload.js -p [ factor-bundle -o add-on/dist/background/background.js -o add-on/dist/options/options.js -o add-on/dist/popup/browser-action.js -o add-on/dist/popup/quick-upload.js ] -o add-on/dist/ipfs-companion-common.js -v",
-    "test": "run-s test:*",
+    "test": "run-s test:functional",
     "test:unit": "cross-env NODE_ENV=test karma start",
     "test:functional": "mocha test/functional/**/*.test.js",
     "lint": "run-s lint:*",
